Remove parsec dependency
Yay, I guess?
