+++ amuletml.cabal
@@ -18,7 +18,6 @@ executable amuletml
                      , base >= 4.9 && < 4.11
                      , lens >= 4.15 && < 4.16
                      , array >= 0.5 && < 0.6
-                     , parsec >= 3.1 && < 3.2
                      , monad-gen >= 0.3 && < 0.4
                      , bytestring >= 0.10.8.2 && < 0.11
                      , containers >= 0.5 && < 0.6
